MAINT: rename v3_api to imopen
