+++ imageio/core/functions.py
@@ -7,7 +7,7 @@ import re
 import numpy as np
 
 from .. import formats
-from .v3_api import imopen
+from .imopen import imopen
 
 
 MEMTEST_DEFAULT_MIM = "256MB"
