+++ imageio/core/imopen.py
